pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.14.0"
requests = "^2.27.1"
python-dateutil = "^2.8.2"
gunicorn = "^20.1.0"
Flask = "^2.1.2"

[tool.poetry.dev-dependencies]
watchdog = "^2.1.8"
pytest = "^7.0.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
